Trait oak_async::ReceiverAsync

pub trait ReceiverAsync {
    type Message: Decodable;
    fn receive_async(&self) -> ChannelRead<Self::Message>;

    fn receive_stream(&self) -> ChannelReadStream<Self::Message> { ... }
}

Extension trait to allow asynchronous reads from a Receiver.

Associated Types

type Message: Decodable

Required methods

fn receive_async(&self) -> ChannelRead<Self::Message>

Asynchronously receive a message.

The returned Future resolves to either a message or an OakError.

Provided methods

fn receive_stream(&self) -> ChannelReadStream<Self::Message>

Asynchronously receive multiple messages.

Each item received from this Stream resolves to either a message or an OakError